Rock state scriptable objects and a better fsm for rocks
