+++ Assets/Prefabs/Rock.prefab
@@ -113,12 +113,12 @@ Rigidbody2D:
   m_UseAutoMass: 0
   m_Mass: 1
   m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
+  m_AngularDrag: 0.53
+  m_GravityScale: 0.76
   m_Material: {fileID: 0}
   m_Interpolate: 0
   m_SleepingMode: 1
-  m_CollisionDetection: 0
+  m_CollisionDetection: 1
   m_Constraints: 0
 --- !u!114 &985767241115572516
 MonoBehaviour:
@@ -132,7 +132,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 1ccfad0e0951541439529fc50f30aa0b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  StateValues:
+  - {fileID: 11400000, guid: b90ac5e7df3ba7147a9220374e94442c, type: 2}
+  - {fileID: 11400000, guid: 232e411c6a0bce1459131ee0539a764a, type: 2}
   isNegative: 0
   rigidbody: {fileID: 985767241115572522}
+  collider: {fileID: 985767241115572523}
   hitWater: 0
-  throwForce: 0.05
+  throwForce: 0.47
+  stateTag:
