+++ Assets/ScriptableObjects/DefaultRockState.asset
@@ -1,0 +1,18 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5e1436dd2d6c832408e973a43d797954, type: 3}
+  m_Name: DefaultRockState
+  m_EditorClassIdentifier: 
+  stateName: default
+  linearDrag: 0
+  angularDrag: 0.54
+  gravityScale: 0.75
