+++ Assets/ScriptableObjects/DefaultRockState.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b90ac5e7df3ba7147a9220374e94442c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
