+++ Assets/ScriptableObjects/SinkingRockState.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 232e411c6a0bce1459131ee0539a764a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
