+++ Assets/Scripts/Rock.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cd5ddb8d34a7bd2449d6f231f9726556
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
